extract getBashCommand, getDenoCommand, use in Installation.
